Migrated all code from the map script into the app to create a final shiny app script (Additionally, realized that I'm terrible at using Git).
Tidied the fuzzyModel Script and added functions to garbage collect.

For the app script - added functions to render basemaps and render pipes coloured by risk.

Added user ability to select different months and rainfall scenarios to change the risk calculation shown.

Also added user ability to visualize the location of pipes based on attributes.

Added a css file to adjust styling on the page.

This is a preliminary draft. Things left to do include tidying the css a little more, editing popups for the basemap and the pipes, and making the overlay between risk and user selection clearer to navigate.